[Refactor] Create Photo.swift to separate code + etc
- PhotoCardsView 안에 들어있던 struct들을 Photo.swift 파일로 옮겨 코드 정리하기
- PhotoCardsView 안 PhotoCard UI 수정
